Move implementation of DeleteVideos to BeatmapModelManager
